Use unified methods to read primitives from Read
This removes a bit of the boilerplate around reading primitives from
std::io::Read.

This trims about 90 lines from the source, making it a bit more
accessible.

The lines in question are also fairly tricky to read and unrelated to
surrounding code, so this helps too.

Alternative
-----------

An alternative would be not to define those methods as a trait extension
for Read, but rather as free functions. Then you would simply call them.

This method is not inferior to the one I chose. But trait extensions are
more rusty.