Add Error impl for TMF error types
When using tmf, I would like to use the `?` operator on Results returned
by tmf methods.

It is currently impossible to do so, because `TMFImportError` and
`TMFExportError` do not implement `Error` (unlike their name suggest).

To fix this, we implement `Error` on TMF types using the thiserror
derive macro.

Alternatives
------------

Instead of using thiserror, it's possible to define yourself the `Error`
implementation. But thiserror is _widely used_ in the rust ecosystem,
and it's fairly tedious (and error-prone) to manually implement `Error`.

There are two drawbacks to using thiserror:

- Compile time are certainly increased (but not by much!) from using a
  proc_macro crate
- It's an additional dependency to worry about

I still think using thiserror is worthwhile. Also considering that it
will probably end up in the user's dependency tree regardless.